Fix another prefix-related problem
This time, there was a problem in listing the things within a
directory.  I had the following directory structure:

* `build`
   * `jenkins-go-package-build`

(in S3 as `build/jenkins-go-package-build/`)

Due to `strings.Replace` being called, the `build/` prefix was
stripped from the front of string (correctly but also from the
back of the string (incorrectly).  This caused any files/folders
that had any part of the prefix in them to be incorrectly displayed
and impossible to navigate to using the directory-listing mechanism.

The resulting directory structure was rendered as:

* `build`
   * `jenkins-go-package-`

This now correctly calls `strings.TrimPrefix`.