Strip repeated slashes from URL path
This will replace the default behaviour of redirecting to serving the
resource directly.
For example for request GET /path//to/file the previous behavior was to
send 301 Moved Permanently with Location: /path/to/file header.
The new behavior is to treat the path as /path/to/file and serve the
file immediately.